extract lines_0_10000.rs
